+++ BaseModel/BaseModel.csproj
@@ -46,12 +46,13 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Polygon.cs" />
     <Compile Include="WorldObjects\Building.cs" />
     <Compile Include="WorldObjects\Iso5218_Person.cs" />
     <Compile Include="WorldObjects\Obstacle.cs" />
     <Compile Include="WorldObjects\ParkingSpot.cs" />
+    <Compile Include="PolygonDictionary.cs" />
     <Compile Include="WorldObjects\Vehicle.cs" />
-    <Compile Include="Polygon.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="WorldObjects\Road.cs" />
     <Compile Include="WorldObjects\RoadSign.cs" />
