+++ BaseModel/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="System.Collections.NonGeneric" version="4.3.0" targetFramework="net472" />
+</packages>
